+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+* Removed support for the `OTEL_SEMCONV_STABILITY_OPT_IN` environment variable
+  which toggled the use of the new conventions for the
+  [server, client, and shared network attributes](https://github.com/open-telemetry/semantic-conventions/blob/v1.23.0/docs/general/attributes.md#server-client-and-shared-network-attributes).
+  Now that this suite of attributes are stable, this instrumentation will only
+  emit the new attributes.
+  ([#5270](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5270))
+* **Breaking Change**: Renamed `SqlClientInstrumentationOptions` to
+  `SqlClientTraceInstrumentationOptions`.
+  ([#5285](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5285))
+
 ## 1.6.0-beta.3
 
 Released 2023-Nov-17